refactor: new structure
