+++ src/psycop_model_evaluation/binary_classification/plots/performance/roc_auc.py
