+++ src/psycop_model_evaluation/binary_classification/plots/robustness/subgroups/performance_by_age.py
@@ -4,14 +4,13 @@ from collections.abc import Sequence
 from pathlib import Path
 from typing import Optional, Union
 
-from psycop_model_training.model_eval.dataclasses import EvalDataset
-
-from psycop_model_evaluation.base_artifacts.plots.base_charts import (
+from psycop_model_evaluation.binary_classification.plots.base_charts import (
     plot_basic_chart,
 )
-from psycop_model_evaluation.base_artifacts.plots.utils import (
+from psycop_model_evaluation.binary_classification.plots.robustness.subgroups.performance_by_input import (
     create_roc_auc_by_input,
 )
+from psycop_model_training.model_eval.dataclasses import EvalDataset
 
 
 def plot_roc_auc_by_age(
