+++ src/psycop_model_evaluation/binary_classification/tables/descriptive_stats_table.py
@@ -5,10 +5,11 @@ from typing import Optional, Union
 
 import pandas as pd
 import wandb
-from psycop_model_training.model_eval.dataclasses import EvalDataset
-
-from psycop_model_evaluation.base_artifacts.tables.tables import output_table
+from psycop_model_evaluation.binary_classification.tables.output_table import (
+    output_table,
+)
 from psycop_model_evaluation.utils.utils import bin_continuous_data
+from psycop_model_training.model_eval.dataclasses import EvalDataset
 
 
 class DescriptiveStatsTable:
