+++ tests/performance_over_calendar_time/test_performance_by_calendar_time.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from psycop_model_evaluation.base_artifacts.plots.performance_over_time import (
+from psycop_model_evaluation.binary_classification.plots.robustness.time.linear.timedelta.data_by_timedelta import (
     create_performance_by_timedelta,
 )
 from psycop_model_training.model_eval.dataclasses import EvalDataset
